+++ doc/WhatsNew.rst
@@ -9,6 +9,7 @@ Ver 5.5.0 (unreleased)
   graphics change in the associated channel viewer
 - fixed an issue where the Preferences plugin did not show the color
   distribution algorithm that was saved in the channel correctly
+- Added internal support for dask and zarr arrays
 
 Ver 5.4.0 (2025-08-10)
 ======================